OC10/requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-nested-routers==0.93.3
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
